@@ -1,29 +1,61 @@
 {
-	"name": "fcc-exercise-tracker",
-	"version": "0.1.0",
-	"description": "A REST API project, part of Free Code Camp's curriculum",
+	"name": "fitness_tracker",
+	"version": "0.0.2",
+	"description": "API project for freeCodeCamp",
 	"main": "server.js",
+	"jest": {
+		"testEnvironment": "node",
+		"setupFiles": [
+			"dotenv/config"
+		],
+		"preset": "@shelf/jest-mongodb"
+	},
 	"scripts": {
-		"start": "node server.js"
+		"build": "rimraf bin/ && babel ./ --out-dir bin/ --ignore ./node_modules,./.babelrc,./package.json,./npm-debug.log --copy-files --source-maps",
+		"start": "npm run build && node -r dotenv/config --inspect bin/server",
+		"start_debug": "npm run build && node -r dotenv/config --inspect bin/server",
+		"test": "jest --detectOpenHandles",
+		"test-url": "jest shorturl-dao.test.js"
 	},
 	"dependencies": {
-		"express": "^4.14.0",
-		"mongoose": "^4.7.2",
-		"mongodb": "^2.2.14",
-		"body-parser": "^1.15.2",
-		"cors": "^2.8.1",
-		"shortid": "^2.2.6"
+		"body-parser": "^1.19.0",
+		"cors": "^2.8.5",
+		"dns": "^0.2.2",
+		"express": "^4.17.1",
+		"mongodb": "^3.3.3",
+		"mongoose": "^5.9.3",
+		"chalk": "^2.4.2",
+		"cookie-parser": "~1.4.4",
+		"debug": "~2.6.9",
+		"express-winston": "^3.1.0",
+		"http-errors": "~1.6.3",
+		"log4js": "^4.1.1",
+		"morgan": "~1.9.1",
+		"winston": "^3.2.1",
+		"@babel/runtime": "^7.8.7"
 	},
 	"engines": {
-		"node": "6.9.1"
+		"node": "10.17.0"
 	},
 	"repository": {
-		"url": "https://gomix.com/#!/project/welcome-project"
+		"type": "git",
+		"url": "https://hyperdev.com/#!/project/welcome-project"
 	},
-	"license": "MIT",
 	"keywords": [
 		"node",
-		"gomix",
-		"express"
-	]
+		"hyperdev",
+		"express",
+		"Babel"
+	],
+	"license": "MIT",
+	"devDependencies": {
+		"@shelf/jest-mongodb": "^1.1.3",
+		"dotenv": "^8.2.0",
+		"jest": "^25.1.0",
+		"@babel/cli": "^7.8.4",
+		"@babel/core": "^7.8.7",
+		"@babel/plugin-transform-runtime": "^7.8.3",
+		"@babel/preset-env": "^7.8.7",
+		"rimraf": "^2.6.3"
+	}
 }
